requirements.txt
torch>=2.0.0
torchvision>=0.15.0
requests>=2.28.0
pillow>=9.0.0
numpy>=1.21.0
python-dotenv>=0.19.0
mercantile>=1.2.0
tqdm>=4.64.0